requirements.txt
openai>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
PyMuPDF>=1.25.0
sentence-transformers>=3.0.0
numpy>=1.26.0
streamlit>=1.38.0
